# web/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_api_words_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          }
        },
        "required": [
          "language",
          "file"
        ],
        "title": "Body_upload_api_words_upload_post",
        "type": "object"
      },
      "DeleteBody": {
        "properties": {
          "ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Ids",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "title": "DeleteBody",
        "type": "object"
      },
      "DeleteResult": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "integer"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "DeleteResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvalidRow": {
        "properties": {
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "row": {
            "title": "Row",
            "type": "integer"
          }
        },
        "required": [
          "row",
          "reason"
        ],
        "title": "InvalidRow",
        "type": "object"
      },
...
      "UploadResult": {
        "properties": {
          "added": {
            "title": "Added",
            "type": "integer"
          },
          "duplicates": {
            "title": "Duplicates",
            "type": "integer"
          },
          "invalid": {
            "items": {
              "$ref": "#/components/schemas/InvalidRow"
            },
            "title": "Invalid",
            "type": "array"
          },
          "invalid_count": {
            "title": "Invalid Count",
            "type": "integer"
          }
        },
        "required": [
          "added",
          "duplicates",
          "invalid_count",
          "invalid"
        ],
        "title": "UploadResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WordIn": {
        "properties": {
          "language": {
            "title": "Language",
            "type": "string"
          },
          "meaning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meaning"
          },
          "word": {
            "title": "Word",
            "type": "string"
          }
        },
        "required": [
          "language",
          "word"
        ],
        "title": "WordIn",
        "type": "object"
      },
      "WordOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "meaning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meaning"
          },
          "word": {
            "title": "Word",
            "type": "string"
          }
        },
        "required": [
          "id",
          "language",
          "word",
          "meaning",
          "created_at"
        ],
        "title": "WordOut",
        "type": "object"
      },
      "WordPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WordOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "WordPage",
        "type": "object"
      },
      "WordPatch": {
        "properties": {
          "meaning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meaning"
          },
          "word": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Word"
          }
        },
        "title": "WordPatch",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Language Tutor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/words": {
      "get": {
        "operationId": "list_words_api_words_get",
        "parameters": [
          {
            "in": "query",
            "name": "language",
            "required": true,
            "schema": {
              "title": "Language",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Words",
        "tags": [
          "words"
        ]
      },
      "post": {
        "operationId": "add_word_api_words_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WordIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Word",
        "tags": [
          "words"
        ]
      }
    },
    "/api/words/delete": {
      "post": {
        "operationId": "delete_words_api_words_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Words",
        "tags": [
          "words"
        ]
      }
    },
    "/api/words/upload": {
      "post": {
        "operationId": "upload_api_words_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_words_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload",
        "tags": [
          "words"
        ]
      }
    },
    "/api/words/{word_id}": {
      "patch": {
        "operationId": "patch_word_api_words__word_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "word_id",
            "required": true,
            "schema": {
              "title": "Word Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WordPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Word",
        "tags": [
          "words"
        ]
      }
    }
  }
}